[4/4] Eng 1802 add mongodb integration - UI (#658)
