+++ src/ui/common/src/components/pages/operator/id/index.tsx
@@ -1,15 +1,7 @@
-import {
-  faCircleDown,
-  faQuestionCircle,
-} from '@fortawesome/free-solid-svg-icons';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { CircularProgress, Divider } from '@mui/material';
 import Box from '@mui/material/Box';
-import IconButton from '@mui/material/IconButton';
-import Tooltip from '@mui/material/Tooltip';
 import Typography from '@mui/material/Typography';
-import { BlobReader, TextWriter, ZipReader } from '@zip.js/zip.js';
-import React, { useEffect, useState } from 'react';
+import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useLocation, useNavigate, useParams } from 'react-router-dom';
 
@@ -18,36 +10,23 @@ import DefaultLayout, {
 } from '../../../../components/layouts/default';
 import { BreadcrumbLink } from '../../../../components/layouts/NavBar';
 import LogViewer from '../../../../components/LogViewer';
-import MultiFileViewer from '../../../../components/MultiFileViewer';
 import { handleGetWorkflowDagResult } from '../../../../handlers/getWorkflowDagResult';
 import { AppDispatch, RootState } from '../../../../stores/store';
-import { theme } from '../../../../styles/theme/theme';
 import UserProfile from '../../../../utils/auth';
 import { getPathPrefix } from '../../../../utils/getPathPrefix';
-import {
-  exportFunction,
-  GoogleSheetsExtractParams,
-  handleExportFunction,
-  hasFile,
-  OperatorType,
-  PREV_TABLE_TAG,
-  RelationalDBExtractParams,
-} from '../../../../utils/operators';
 import {
   isInitial,
   isLoading,
   LoadingStatusEnum,
 } from '../../../../utils/shared';
-import { CodeBlock } from '../../../CodeBlock';
-import { Button } from '../../../primitives/Button.styles';
 import ArtifactSummaryList from '../../../workflows/artifact/summaryList';
+import OperatorSpecDetails from '../../../workflows/operator/specDetails';
 import DetailsPageHeader from '../../components/DetailsPageHeader';
 import { LayoutProps } from '../../types';
 
 type OperatorDetailsPageProps = {
   user: UserProfile;
   Layout?: React.FC<LayoutProps>;
-  maxRenderSize?: number;
   workflowIdProp?: string;
   workflowDagResultIdProp?: string;
   operatorIdProp?: string;
@@ -58,7 +37,6 @@ type OperatorDetailsPageProps = {
 const OperatorDetailsPage: React.FC<OperatorDetailsPageProps> = ({
   user,
   Layout = DefaultLayout,
-  maxRenderSize = 100000000,
   workflowIdProp,
   workflowDagResultIdProp,
   operatorIdProp,
@@ -80,14 +58,6 @@ const OperatorDetailsPage: React.FC<OperatorDetailsPageProps> = ({
   if (operatorIdProp) {
     operatorId = operatorIdProp;
   }
-
-  const [files, setFiles] = useState({
-    '': {
-      path: '',
-      language: 'plaintext',
-      content: '',
-    },
-  });
 
   const workflowDagResultWithLoadingStatus = useSelector(
     (state: RootState) =>
@@ -142,68 +112,6 @@ const OperatorDetailsPage: React.FC<OperatorDetailsPageProps> = ({
   const logs = operator?.result?.exec_state?.user_logs ?? {};
   const operatorError = operator?.result?.exec_state?.error;
 
-  const setFileHelper = (prevState, file, fileContents) => {
-    const nextState = { ...prevState };
-    const pathList = file.filename.split('/');
-    let base = nextState;
-    pathList.forEach((section, i) => {
-      // Create a key for each first-level subfolder
-      if (!Object.keys(base).includes(section)) {
-        base[section] = {};
-      }
-      if (!file.directory && i + 1 === pathList.length) {
-        // Include the file metadata
-        base[section] = fileContents;
-      } else {
-        // Go into the subfolder
-        base = base[section];
-      }
-    });
-    return nextState;
-  };
-
-  useEffect(() => {
-    async function getFilesBlob() {
-      // This is the function used to retrieve the contents in the function that generates the operator's zip file.
-      const blob = await exportFunction(user, operatorId);
-      if (blob) {
-        const reader = new ZipReader(new BlobReader(blob));
-        const entries = await reader.getEntries();
-        entries.forEach((file) => {
-          let language = 'plaintext';
-          if (file.filename.endsWith('.py')) {
-            language = 'python';
-          }
-          if (file.uncompressedSize < maxRenderSize) {
-            file.getData(new TextWriter()).then((content) => {
-              setFiles((prevState) =>
-                setFileHelper(prevState, file, {
-                  path: file.filename,
-                  language: language,
-                  content: content,
-                })
-              );
-            });
-          } else {
-            setFiles((prevState) =>
-              setFileHelper(prevState, file, {
-                path: file.filename,
-                language: 'plaintext',
-                content:
-                  'We do not support viewing such large files.\nPlease download this file instead.',
-              })
-            );
-          }
-        });
-        await reader.close();
-      }
-    }
-
-    if (hasFile(operator?.spec?.type)) {
-      getFilesBlob();
-    }
-  }, [maxRenderSize, operator, operatorId, user]);
-
   if (
     !workflowDagResultWithLoadingStatus ||
     isInitial(workflowDagResultWithLoadingStatus.status) ||
@@ -236,116 +144,6 @@ const OperatorDetailsPage: React.FC<OperatorDetailsPageProps> = ({
       .filter((artf) => !!artf);
   const inputs = mapArtifacts(operator.inputs);
   const outputs = mapArtifacts(operator.outputs);
-
-  const operatorPreview = () => {
-    if (hasFile(operator?.spec?.type)) {
-      return (
-        <Box>
-          <Box mb={1} display="flex" flexDirection="row">
-            <Typography variant="h6" fontWeight="normal">
-              Code Preview
-            </Typography>
-            <Box flexGrow={1} />
-            <Button
-              onClick={async () => {
-                await handleExportFunction(
-                  user,
-                  operator.id,
-                  `${operator.name}.zip`
-                );
-              }}
-              color="secondary"
-            >
-              <FontAwesomeIcon icon={faCircleDown} />
-              <Typography sx={{ ml: 1 }}>{`${operator.name}.zip`}</Typography>
-            </Button>
-          </Box>
-          <MultiFileViewer files={files} defaultFile={operator.name || ''} />
-        </Box>
-      );
-    }
-
-    if (operator?.spec?.type === OperatorType.Extract) {
-      const extractParams = operator.spec.extract.parameters;
-      let content = null;
-
-      if ('query' in extractParams || 'queries' in extractParams) {
-        // relational
-        const relationalParams = extractParams as RelationalDBExtractParams;
-        const renderQuery = (q: string) => (
-          <CodeBlock language="sql">{q}</CodeBlock>
-        );
-        let tooltips = '';
-        const chainTagTooltips =
-          '`$` refers to the output of the previous query.';
-
-        if (!!relationalParams.queries && relationalParams.queries.length > 0) {
-          const queries = relationalParams.queries;
-          content = (
-            <Box display="flex" flexDirection="column">
-              {relationalParams.queries.map((q, idx) => (
-                <Box mb={1} key={`extract-query-${idx}`}>
-                  {renderQuery(q)}
-                </Box>
-              ))}
-            </Box>
-          );
-          const hasChainTag = queries.some((q) => q.includes(PREV_TABLE_TAG));
-          tooltips = `These queries are chained. ${
-            hasChainTag ? chainTagTooltips : ''
-          }`;
-        } else {
-          content = renderQuery(relationalParams.query);
-        }
-
-        return (
-          <Box>
-            <Box display="flex" flexDirection="row" marginBottom={1}>
-              <Typography
-                variant="h6"
-                fontWeight="normal"
-                alignContent="center"
-              >
-                Query Details
-              </Typography>
-              {tooltips && (
-                <Tooltip arrow placement="right" title={tooltips}>
-                  <IconButton>
-                    <FontAwesomeIcon
-                      color={`${theme.palette.gray[700]}`}
-                      fontSize="16px"
-                      icon={faQuestionCircle}
-                    />
-                  </IconButton>
-                </Tooltip>
-              )}
-            </Box>
-            {content}
-          </Box>
-        );
-      }
-
-      // google sheet
-      const googleSheetsParams = extractParams as GoogleSheetsExtractParams;
-      return (
-        <Box>
-          <Typography variant="h6" fontWeight="normal" mb={1}>
-            Spreadsheet Details
-          </Typography>
-          <Typography variant="body1" mb={1}>
-            <strong>Spreadsheet ID: </strong>
-            {googleSheetsParams.spreadsheet_id}
-          </Typography>
-          <Typography variant="body1">
-            <strong>Query: </strong>
-            {googleSheetsParams.query}
-          </Typography>
-        </Box>
-      );
-    }
-
-    return null;
-  };
 
   return (
     <Layout breadcrumbs={breadcrumbs} user={user}>
@@ -395,7 +193,7 @@ const OperatorDetailsPage: React.FC<OperatorDetailsPageProps> = ({
           </Box>
 
           <Divider sx={{ my: '32px' }} />
-          {operatorPreview()}
+          <OperatorSpecDetails user={user} operator={operator} />
         </Box>
       </Box>
     </Layout>
