+++ src/ui/common/src/utils/integrations.ts
@@ -61,6 +61,11 @@ export type MariaDbConfig = {
   password?: string;
 };
 
+export type MongoDBConfig = {
+  auth_uri: string;
+  database: string;
+};
+
 export type SqlServerConfig = {
   host: string;
   port: string;
@@ -177,7 +182,8 @@ export type Service =
   | 'Kubernetes'
   | 'SQLite'
   | 'Lambda'
-  | 'Google Sheets';
+  | 'Google Sheets'
+  | 'MongoDB';
 
 export type Info = {
   logo: string;
@@ -338,6 +344,11 @@ export const SupportedIntegrations: ServiceInfoMap = {
     logo: 'https://aqueduct-public-assets-bucket.s3.us-east-2.amazonaws.com/webapp/pages/integrations/Lambda.png',
     activated: true,
     category: 'compute',
+  },
+  ['MongoDB']: {
+    logo: 'https://aqueduct-public-assets-bucket.s3.us-east-2.amazonaws.com/webapp/pages/integrations/mongo.png',
+    activated: true,
+    category: 'data',
   },
 };
